Reorganise code structure for a more functionality
