+++ syslib/src/blk.rs
