+++ syslib/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod blk;
 pub mod fs;
-pub mod kmodprobe;
